done for 5/11

+++ venueSearch/venueSearch/Event.swift
@@ -17,7 +17,7 @@ class Event : NSManagedObject {
     @NSManaged var start: [String : AnyObject]
     @NSManaged var popularity: Double
     @NSManaged var location: [String: AnyObject]
-    @NSManaged var id: Int
+    @NSManaged var id: Int64
     @NSManaged var performance: [[String : AnyObject]]
     @NSManaged var venue: Venue?
     
@@ -37,7 +37,7 @@ class Event : NSManagedObject {
         start = dictionary[SongKickClient.JSONResponseKeys.Start] as! [String : AnyObject]
         popularity = dictionary[SongKickClient.JSONResponseKeys.Popularity] as! Double
         location = dictionary[SongKickClient.JSONResponseKeys.Location] as! [String: AnyObject]
-        id = dictionary[SongKickClient.JSONResponseKeys.ID] as! Int
+        id = dictionary[SongKickClient.JSONResponseKeys.ID] as! Int64
         performance = dictionary[SongKickClient.JSONResponseKeys.Performance] as! [[String : AnyObject]]
         venue = dictionary[SongKickClient.JSONResponseKeys.Venue] as? Venue
     }
